Address review feedback for PR #948
- Revert localization-fetcher deepMerge change; restore the original
  cached -> bundled -> en fallback chain.
- Deduplicate updateLast and updateSpecific cases in anki.ts via a
  shared private helper, _updateNoteFields.
- Drop the separate CardSelectUi component, controller, entrypoint
  and iframe; the card selector is now another mode of AnkiUi so all
  the bridge, settings and Anki client wiring is reused.
- Extract the picker into a reusable component,
  common/components/CardSelectView.tsx, so the webapp can pick it up
  later.
